refactor : move excp pakage
