+++ backend/src/main/java/toonpick/app/service/AuthorService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import toonpick.app.dto.AuthorDTO;
 import toonpick.app.domain.webtoon.Author;
 import toonpick.app.exception.ErrorCode;
-import toonpick.app.exception.ResourceAlreadyExistsException;
-import toonpick.app.exception.ResourceNotFoundException;
+import toonpick.app.exception.exception.ResourceAlreadyExistsException;
+import toonpick.app.exception.exception.ResourceNotFoundException;
 import toonpick.app.mapper.AuthorMapper;
 import toonpick.app.repository.AuthorRepository;
 import org.springframework.stereotype.Service;
