+++ backend/src/main/java/toonpick/app/service/FavoriteToonService.java
@@ -8,8 +8,8 @@ import toonpick.app.domain.member.Member;
 import toonpick.app.domain.webtoon.Webtoon;
 import toonpick.app.dto.webtoon.WebtoonResponseDTO;
 import toonpick.app.exception.ErrorCode;
-import toonpick.app.exception.ResourceAlreadyExistsException;
-import toonpick.app.exception.ResourceNotFoundException;
+import toonpick.app.exception.exception.ResourceAlreadyExistsException;
+import toonpick.app.exception.exception.ResourceNotFoundException;
 import toonpick.app.mapper.WebtoonMapper;
 import toonpick.app.repository.FavoriteToonRepository;
 import toonpick.app.repository.MemberRepository;
