+++ backend/src/main/java/toonpick/app/service/GenreService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import toonpick.app.dto.GenreDTO;
 import toonpick.app.domain.webtoon.Genre;
 import toonpick.app.exception.ErrorCode;
-import toonpick.app.exception.ResourceAlreadyExistsException;
-import toonpick.app.exception.ResourceNotFoundException;
+import toonpick.app.exception.exception.ResourceAlreadyExistsException;
+import toonpick.app.exception.exception.ResourceNotFoundException;
 import toonpick.app.mapper.GenreMapper;
 import toonpick.app.repository.GenreRepository;
 import org.springframework.stereotype.Service;
