+++ backend/src/main/java/toonpick/app/service/WebtoonReviewService.java
@@ -16,9 +16,9 @@ import toonpick.app.domain.review.ReviewLike;
 import toonpick.app.domain.member.Member;
 import toonpick.app.domain.webtoon.Webtoon;
 import toonpick.app.domain.review.WebtoonReview;
-import toonpick.app.exception.DatabaseOperationException;
+import toonpick.app.exception.exception.DatabaseOperationException;
 import toonpick.app.exception.ErrorCode;
-import toonpick.app.exception.ResourceNotFoundException;
+import toonpick.app.exception.exception.ResourceNotFoundException;
 import toonpick.app.mapper.WebtoonReviewMapper;
 import toonpick.app.repository.ReviewLikeRepository;
 import toonpick.app.repository.MemberRepository;
